+++ lib/chef-dk/command/undelete.rb
@@ -46,7 +46,7 @@ manually reapply any ACL customizations you have made.
 
 See our detailed README for more information:
 
-https://github.com/chef/chef-dk/blob/master/POLICYFILE_README.md
+https://docs.chef.io/policyfile.html
 
 Options: