+++ lib/chef-dk/command/update.rb
@@ -42,7 +42,7 @@ future version.
 
 See our detailed README for more information:
 
-https://github.com/chef/chef-dk/blob/master/POLICYFILE_README.md
+https://docs.chef.io/policyfile.html
 
 Options: