+++ omnibus/config/software/chef-dk-cleanup.rb
@@ -1,5 +1,5 @@
 #
-# Copyright 2012-2014 Chef Software, Inc.
+# Copyright 2016 Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-name "chef-dk-remove-docs"
+name "chef-dk-cleanup"
 
 license :project_license
 
@@ -23,6 +23,14 @@ build do
   require_relative "../../files/chef-dk/build-chef-dk"
   extend BuildChefDK
 
+  # Clear the now-unnecessary git caches, cached gems, and git-checked-out gems
+  block "Delete bundler git cache and git installs" do
+    gemdir = shellout!("#{gem_bin} environment gemdir", env: env).stdout.chomp
+    remove_directory "#{gemdir}/cache"
+    remove_directory "#{gemdir}/bundler"
+  end
+
+  # Clean up docs
   delete "#{install_dir}/embedded/docs"
   delete "#{install_dir}/embedded/share/man"
   delete "#{install_dir}/embedded/share/doc"